ref: UI improvements
Ref #55
