+++ .gitignore
@@ -152,3 +152,7 @@ app.*.map.json
 !.idea/codeStyles/
 !.idea/dictionaries/
 !.idea/runConfigurations/
+
+# Custom ignores
+
+wiki/
