+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "packages/a2f_sdk"]
 	path = packages/a2f_sdk
 	url = https://github.com/app-anvil/a2f_sdk.git
+[submodule "wiki"]
+	path = wiki
+	url = https://github.com/app-anvil/waste_not.wiki.git
