+++ README.md
@@ -1,0 +1,37 @@
+# Waste NOT! - Your food waste reduction companion 
+  
+[![GitHub Release](https://img.shields.io/github/v/release/app-anvil/waste_not)](https://github.com/app-anvil/a2f_sdk/releases)
+[![License](https://img.shields.io/github/license/app-anvil/waste_not)](https://opensource.org/licenses/MIT)
+[![GitHub Issues](https://img.shields.io/github/issues/app-anvil/waste_not.svg)](https://github.com/app-anvil/waste_not/issues)
+[![GitHub Pull Requests](https://img.shields.io/github/issues-pr/app-anvil/waste_not.svg)](https://github.com/app-anvil/waste_not/pulls)
+
+A food waste reduction app that helps users track their food inventory and expiration dates.
+
+## Table of Contents
+
+- [Useful links](#useful-links)
+- [Installation](#installation)
+- [Usage](#usage)
+- [Contributing](#contributing)
+
+## Useful links
+
+- [Wiki](wiki)
+
+## Installation
+
+Instructions on how to install and run the project.
+
+## Usage
+
+Instructions on how to use the project.
+
+## Contributing
+
+Closed.
+
+## Structure
+
+### App
+
+- `app`: Contains the main application code.
