+++ analysis_options.yaml
@@ -1,43 +1,1 @@
-include: package:very_good_analysis/analysis_options.yaml
-analyzer:
-  exclude:
-    [
-      build/**,
-      lib/**.freezed.dart,
-      packages/**.freezed.dart,
-      lib/**.g.dart,
-      packages/**.g.dart,
-    ]
-linter:
-  rules:
-    always_use_package_imports: false
-    always_declare_return_types: true
-    avoid_bool_literals_in_conditional_expressions: true
-    avoid_print: true
-    avoid_redundant_argument_values: false
-    avoid_unused_constructor_parameters: true
-    avoid_void_async: true
-    await_only_futures: true
-    # depend_on_referenced_packages: false
-    flutter_style_todos: false
-    # library_private_types_in_public_api: false
-    lines_longer_than_80_chars: true
-    prefer_const_constructors: true
-    prefer_const_declarations: true
-    prefer_const_literals_to_create_immutables: true
-    prefer_final_in_for_each: true
-    prefer_final_locals: true
-    prefer_function_declarations_over_variables: false
-    prefer_interpolation_to_compose_strings: false
-    prefer_relative_imports: true
-    public_member_api_docs: false
-    sort_constructors_first: false
-    unawaited_futures: true
-    unnecessary_await_in_return: true
-    unrelated_type_equality_checks: true
-    # use_build_context_synchronously: false
-    use_if_null_to_convert_nulls_to_bools: true
-    use_named_constants: true
-    prefer_if_elements_to_conditional_expressions: false
-    comment_references: false
-    overridden_fields: false
+include: package:a2f_sdk/analysis_options.yaml
