+++ app/lib/boot/bootstrap.dart
@@ -1,0 +1,18 @@
+import 'dart:async';
+import 'dart:developer';
+
+import 'package:flutter/widgets.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import 'app_bloc_observer.dart';
+
+Future<void> bootstrap(FutureOr<Widget> Function() builder) async {
+  FlutterError.onError = (details) {
+    log(details.exceptionAsString(), stackTrace: details.stack);
+  };
+
+  Bloc.observer = const AppBlocObserver();
+
+  // Add cross-flavor configuration here
+  runApp(await builder());
+}
