+++ app/lib/features/storages/view/storages_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:storages_repository/storages_repository.dart';
 
-import '../../../routes/app_route.dart';
+import '../../../router/app_route.dart';
 import '../storages.dart';
 
 class StoragesPage extends StatelessWidget {
