+++ app/lib/main_production.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/widgets.dart';
 
-import 'app/app.dart';
-import 'bootstrap.dart';
-import 'injection.dart';
+import 'app/waste_not_app.dart';
+import 'boot/bootstrap.dart';
+import 'boot/get_it.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   initializeDependencies();
-  bootstrap(() => const App());
+  bootstrap(() => const WasteNotApp());
 }
