+++ app/lib/router/app_route.dart
@@ -10,7 +10,6 @@ enum AppRoute {
   storages,
   addStorage,
   editStorage,
-
   editItem,
 }
